Move dataframe commun API in lale.lib
